requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fakeredis==2.27.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mongomock==4.3.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
sentinels==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
